Nach Update auf GCC 10.3
-fcommon ergänzt, sonst hat der Linker alles mehrfach genommen.

Kommentar von
"https://lists.fedoraproject.org/archives/list/devel@lists.fedoraproject.org/thread/RYVPP45PMWPXYBBDKXO3CI7YGL7CDQG6/":

Default to -fno-common
A common mistake in C is omitting extern when declaring a global
variable in a
header file. If the header is included by several files it results in
multiple
definitions of the same variable. In previous GCC versions this error is
ignored. GCC 10 defaults to -fno-common, which means a linker error will
now be
reported. To fix this, use extern in header files when declaring global
variables, and ensure each global is defined in exactly one C file. As a
workaround, legacy C code can be compiled with -fcommon.

+++ USB_DEVICE/App/usbd_cdc_if.c
@@ -439,10 +439,10 @@ static int8_t CDC_Receive_FS(uint8_t* Buf, uint32_t *Len)
   * @param  Len: Number of data to be sent (in bytes)
   * @retval USBD_OK if all operations are OK else USBD_FAIL or USBD_BUSY
   */
-uint8_t CDC_Transmit_FS(uint8_t* Buf, size_t Len)
+uint8_t CDC_Transmit_FS(uint8_t* Buf, uint16_t Len)
 {
-	uint8_t result = USBD_OK;
-	/* USER CODE BEGIN 7 */
+  uint8_t result = USBD_OK;
+  /* USER CODE BEGIN 7 */
 	// aus: https://nefastor.com/microcontrollers/stm32/usb/stm32cube-usb-device-library/communication-device-class/
 
 	//Step 1 : calculate the occupied space in the Tx FIFO
